214-A1/requirements.txt
colorama==0.4.6
PyJWT==2.8.0
